[APM] Migrate test `apm-api-integration/tests/errors` to be deployment agnostic api tests (#199655)
Closes #198971
Part of #193245

## Summary

This PR migrates test apm-api-integration/tests/errors to be deployment
agnostic api tests

## Testing

Serverless:

```
node scripts/functional_tests_server --config x-pack/test/api_integration/deployment_agnostic/configs/serverless/oblt.serverless.config.ts
node scripts/functional_test_runner --config x-pack/test/api_integration/deployment_agnostic/configs/serverless/oblt.serverless.config.ts --grep="APM"
```
Stateful:

```
node scripts/functional_tests_server --config x-pack/test/api_integration/deployment_agnostic/configs/stateful/oblt.stateful.config.ts
node scripts/functional_test_runner --config x-pack/test/api_integration/deployment_agnostic/configs/stateful/oblt.stateful.config.ts --grep="APM"
```

(cherry picked from commit 7921f5618d6af8e959ed4c52a91a026c672955ea)
