+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/errors/top_erroneous_transactions/generate_data.ts
@@ -8,15 +8,15 @@ import { apm, timerange } from '@kbn/apm-synthtrace-client';
 import type { ApmSynthtraceEsClient } from '@kbn/apm-synthtrace';
 
 export const config = {
-  firstTransaction: {
+  appleTransaction: {
     name: 'GET /apple 🍎',
-    successRate: 75,
-    failureRate: 25,
+    successRate: 25,
+    failureRate: 75,
   },
-  secondTransaction: {
+  bananaTransaction: {
     name: 'GET /banana 🍌',
-    successRate: 50,
-    failureRate: 50,
+    successRate: 80,
+    failureRate: 20,
   },
 };
 
@@ -35,14 +35,12 @@ export async function generateData({
     .service({ name: serviceName, environment: 'production', agentName: 'go' })
     .instance('instance-a');
 
-  const interval = '1m';
+  const { bananaTransaction, appleTransaction } = config;
+  const interval = timerange(start, end).interval('1m');
 
-  const { firstTransaction, secondTransaction } = config;
-
-  const documents = [firstTransaction, secondTransaction].flatMap((transaction, index) => {
+  const documents = [appleTransaction, bananaTransaction].flatMap((transaction, index) => {
     return [
-      timerange(start, end)
-        .interval(interval)
+      interval
         .rate(transaction.successRate)
         .generator((timestamp) =>
           serviceGoProdInstance
@@ -51,24 +49,21 @@ export async function generateData({
             .duration(1000)
             .success()
         ),
-      timerange(start, end)
-        .interval(interval)
-        .rate(transaction.failureRate)
-        .generator((timestamp) =>
-          serviceGoProdInstance
-            .transaction({ transactionName: transaction.name })
-            .errors(
-              serviceGoProdInstance
-                .error({ message: `Error 1 transaction ${transaction.name}` })
-                .timestamp(timestamp),
-              serviceGoProdInstance
-                .error({ message: `Error 2 transaction ${transaction.name}` })
-                .timestamp(timestamp)
-            )
-            .duration(1000)
-            .timestamp(timestamp)
-            .failure()
-        ),
+      interval.rate(transaction.failureRate).generator((timestamp) =>
+        serviceGoProdInstance
+          .transaction({ transactionName: transaction.name })
+          .errors(
+            serviceGoProdInstance
+              .error({ message: `Error 1`, type: transaction.name })
+              .timestamp(timestamp),
+            serviceGoProdInstance
+              .error({ message: `Error 2`, type: transaction.name })
+              .timestamp(timestamp)
+          )
+          .duration(1000)
+          .timestamp(timestamp)
+          .failure()
+      ),
     ];
   });
