+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/index.ts
@@ -12,6 +12,7 @@ export default function apmApiIntegrationTests({
 }: DeploymentAgnosticFtrProviderContext) {
   describe('APM', function () {
     loadTestFile(require.resolve('./agent_explorer'));
+    loadTestFile(require.resolve('./errors'));
     loadTestFile(require.resolve('./alerts'));
     loadTestFile(require.resolve('./custom_dashboards'));
     loadTestFile(require.resolve('./dependencies'));
